requirements.txt
Flask
Gunicorn
Flask-SQLAlchemy
Flask-Login
intasend-python
Flask-WTF
email-validator